airflow-docker/requirements.txt
google-cloud-bigquery
pandas
requests
apache-airflow-providers-google==10.12.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
